fix(repository): altera parâmetros de métodos para receber o id do usuário

+++ src/Model/Repository/TaskRepository.php
@@ -14,35 +14,39 @@ class TaskRepository
 
     public function add(Task $task): bool
     {
-        $sql = "INSERT INTO tasks (title, description, task_date, status) VALUES (:title, :description, :task_date, :status)";
+        $sql = "INSERT INTO tasks (title, description, task_date, user_id, status) VALUES (:title, :description, :task_date, :user_id, :status)";
         $statement = $this->pdo->prepare($sql);
         $statement->bindValue(":title", $task->title);
         $statement->bindValue(":description", $task->description);
         $statement->bindValue(":task_date", $task->date);
+		$statement->bindValue(":user_id", $task->userId, PDO::PARAM_INT);
 		$statement->bindValue(":status", $task->getStatus()->value);
         return $statement->execute();
     }
 
     public function edit(Task $task): bool
     {
-        $sql = "UPDATE tasks SET title = :title, description = :description, task_date = :task_date, status = :status WHERE id = :id";
+        $sql = "UPDATE tasks SET title = :title, description = :description, task_date = :task_date, status = :status WHERE id = :id AND user_id = :user_id";
         $statement = $this->pdo->prepare($sql);
         return $statement->execute([
             ":title" => $task->title,
             ":description" => $task->description,
             ":task_date" => $task->date,
             ":status" => $task->getStatus()->value,
-            ":id" => $task->id
+            ":id" => $task->id,
+			":user_id" => $task->userId
         ]);
     }
 
     /**
      * @return Task[]
      */
-    public function all(): array
+    public function all(int $userId): array
     {
-        $sql = "SELECT * FROM tasks";
-        $statement = $this->pdo->query($sql);
+        $sql = "SELECT * FROM tasks WHERE user_id = ?";
+        $statement = $this->pdo->prepare($sql);
+		$statement->bindValue(1, $userId, PDO::PARAM_INT);
+		$statement->execute();
         $data = $statement->fetchAll(PDO::FETCH_ASSOC);
 
 		$tasks = [];
@@ -54,6 +58,7 @@ class TaskRepository
 				$row['title'],
 				$row['description'],
 				$row['task_date'],
+				$row['user_id'],
 				$status
 			);
 
@@ -67,11 +72,12 @@ class TaskRepository
     /**
      * @return null|Task
      */
-    public function find(int $id): ?Task
+    public function find(int $id, int $userId): ?Task
     {
-        $sql = "SELECT * FROM tasks WHERE id = ?";
+        $sql = "SELECT * FROM tasks WHERE id = ? AND user_id = ?";
         $statement = $this->pdo->prepare($sql);
         $statement->bindValue(1, $id, PDO::PARAM_INT);
+        $statement->bindValue(2, $userId, PDO::PARAM_INT);
         if($statement->execute() === false) {
             return null;
         }
@@ -87,6 +93,7 @@ class TaskRepository
 			$row['title'],
 			$row['description'],
 			$row['task_date'],
+			$row['user_id'],
 			$status
 		);
 
@@ -95,20 +102,22 @@ class TaskRepository
 		return $task;
 	}
 
-    public function remove(int $id): bool
+    public function remove(int $id, int $userId): bool
     {
-        $sql = "DELETE FROM tasks WHERE id = ?";
+        $sql = "DELETE FROM tasks WHERE id = ? AND user_id = ?";
         $statement = $this->pdo->prepare($sql);
         $statement->bindValue(1, $id, PDO::PARAM_INT);
+        $statement->bindValue(2, $userId, PDO::PARAM_INT);
         return $statement->execute();
     }
 
-	public function conclude(int $id): bool
+	public function conclude(int $id, int $userId): bool
 	{
-		$sql = "UPDATE tasks SET status = :concluded WHERE id = :id";
+		$sql = "UPDATE tasks SET status = :concluded WHERE id = :id AND user_id = :user_id";
 		$statement = $this->pdo->prepare($sql);
 		$statement->bindValue(":concluded", Status::Concluded->value);
 		$statement->bindValue(":id", $id, PDO::PARAM_INT);
+		$statement->bindValue(":user_id", $userId, PDO::PARAM_INT);
 		return $statement->execute();
 	}
 
@@ -125,6 +134,7 @@ class TaskRepository
 				$row['title'],
 				$row['description'],
 				$row['task_date'],
+				$row['user_id'],
 				$status
 			);
 			$task->setId($row['id']);
@@ -140,5 +150,16 @@ class TaskRepository
 		$statement->bindValue(":expired", Status::Expired->value);
 		$statement->bindValue(":pending", Status::Pending->value);
 		$statement->execute();
+	}
+
+	public function resetTask(int $id, int $userId): bool
+	{
+		$sql = "UPDATE tasks SET status = ? WHERE id = ? AND user_id = ?";
+		$statement = $this->pdo->prepare($sql);
+		return $statement->execute([
+			Status::Pending->value,
+			$id,
+			$userId
+		]);
 	}
 }
